String prototype: split into separate files
